Fit_Plan_Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib==1.7.4
bcrypt==4.0.1
python-dotenv
email-validator
